Basic Android file support

+++ LocalJournal/LocalJournal/Services/FileDataStore.cs
@@ -9,31 +9,41 @@ namespace LocalJournal.Services
 {
 	public class FileDataStore : IDataStore<TextEntry>
 	{
-		public static string DataPath = Path.Combine(FileSystem.AppDataDirectory, "journal");
+		public string DataPath { get; protected set; }
 
 		private readonly IDataSerializer<TextEntry> dataSerializer;
 
 		public FileDataStore()
 		{
 			dataSerializer = DependencyService.Get<IDataSerializer<TextEntry>>();
+			DataPath = Path.Combine(FileSystem.AppDataDirectory, "journal");
 		}
 
 		public async Task<bool> AddEntryAsync(TextEntry entry)
 		{
+			if (!CheckPermission())
+				return false;
+
+			Directory.CreateDirectory(DataPath);
 			var file = FileFromId(entry.Id);
-			Directory.CreateDirectory(file);
 			using (var sw = new StreamWriter(file))
 				return await dataSerializer.WriteAsync(sw, entry);
 		}
 
 		public async Task<bool> UpdateEntryAsync(TextEntry entry)
 		{
+			if (!CheckPermission())
+				return false;
+
 			using (var sw = new StreamWriter(FileFromId(entry.Id)))
 				return await dataSerializer.WriteAsync(sw, entry);
 		}
 
 		public async Task<bool> DeleteEntryAsync(string id)
 		{
+			if (!CheckPermission())
+				return false;
+
 			File.Delete(FileFromId(id));
 
 			return await Task.FromResult(true);
@@ -41,12 +51,18 @@ namespace LocalJournal.Services
 
 		public async Task<TextEntry> GetEntryAsync(string id)
 		{
+			if (!CheckPermission())
+				return null;
+
 			using (var sr = new StreamReader(FileFromId(id)))
 				return await dataSerializer.ReadAsync(sr, id);
 		}
 
 		public async Task<IEnumerable<TextEntry>> GetEntriesAsync(bool forceRefresh = false)
 		{
+			if (!CheckPermission())
+				return new List<TextEntry>(0);
+
 			var files = Directory.GetFiles(DataPath, "*.md");
 			var entries = new List<TextEntry>(files.Length);
 			foreach (var file in files)
@@ -58,9 +74,14 @@ namespace LocalJournal.Services
 			return await Task.FromResult(entries);
 		}
 
-		private string FileFromId(string id)
+		protected virtual string FileFromId(string id)
 		{
 			return Path.Combine(DataPath, $"{id}.md");
+		}
+
+		protected virtual bool CheckPermission()
+		{
+			return true;
 		}
 	}
 }
